Moved tests under library to avoid name collision with mvtest
