+++ libgwas/tests/bedfiles/analysis.bed
