+++ libgwas/tests/bedfiles/analysis.bim
