+++ libgwas/tests/bedfiles/analysis.fam
