+++ libgwas/tests/bedfiles/analysis.vcf
