+++ libgwas/tests/bedfiles/analysis.vcf.gz
