+++ libgwas/tests/bedfiles/bgen_nomiss.sample
