+++ libgwas/tests/bedfiles/miss.sample
