+++ libgwas/tests/bedfiles/miss.vcf
