+++ libgwas/tests/bedfiles/miss.vcf.gz
