+++ libgwas/tests/bedfiles/miss.vcf.gz.tbi
