+++ libgwas/tests/bedfiles/nomiss.sample
