+++ libgwas/tests/bedfiles/nomiss.vcf
