+++ libgwas/tests/bedfiles/nomiss.vcf.gz
