+++ libgwas/tests/bedfiles/nomiss.vcf.gz.tbi
