+++ libgwas/tests/bedfiles/ped_missing.bed
