+++ libgwas/tests/bedfiles/ped_missing.bim
