+++ libgwas/tests/bedfiles/ped_nomiss.bed
