+++ libgwas/tests/bedfiles/ped_nomiss.bim
