+++ libgwas/tests/bedfiles/ped_nomiss.fam
