+++ libgwas/tests/bedfiles/test.bgen
