+++ libgwas/tests/bedfiles/test.bgen.metadata
