+++ libgwas/tests/bedfiles/test.bgen.sample
