+++ libgwas/tests/test_transped_parser.py
