+++ setup.py
@@ -20,14 +20,14 @@ setuptools.setup(name="libgwas",
     author_email="eric.s.torstenson@vanderbilt.edu",
     url="https://github.com/edwards-lab/libGWAS",
     download_url="https://github.com/edwards-lab/libGWAS/archive/v1.1.0.tar.gz",
-    packages=["libgwas", "tests"],
+    packages=["libgwas", "libgwas.tests"],
     license="GPL",
     description=["GWAS Parser Library"],
     install_requires=["scipy","numpy"],
     long_description=read('README'),
     keywords=["GWAS","genetic analysis"],
-    test_suite='tests',
-    package_data={'tests/bedfiles/':['*'],
+    test_suite='libgwas.tests',
+    package_data={'libgwas/tests/bedfiles/':['*'],
                   'doc':['*']},
     classifiers=[
         "Development Status :: 4 - Beta",
